Добавление возможностей фильтрации, поиска и сортировки данных

+++ Suppliers.Backend/Suppliers.Application/Products/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Suppliers.Application.Common.Exceptions;
 using Suppliers.Application.Interfaces;
 using Suppliers.Domain;
@@ -18,6 +19,13 @@ namespace Suppliers.Application.Products.Commands.DeleteProduct
             if (entity == null)
             {
                 throw new NotFoundException(nameof(Product), request.Id);
+            }
+
+            var batches = await _context.Batches.Include(batch => batch.Product).ToListAsync(cancellationToken);
+
+            if(batches.FirstOrDefault(batch => batch.ProductId == entity.Id) != null)
+            {
+                throw new Exception("Невозможно удалить информацию о товаре, для которого существует активный заказ");
             }
 
             _context.Products.Remove(entity);
